requirements.txt
# ── Core Azure & Graph ─────────────────────────────────────────────────────
azure-identity>=1.13.0
msgraph-sdk>=0.2.0
azure-keyvault-secrets>=4.4.0
azure-search-documents>=11.4.0
azure-core>=1.28.0

# ── LLM Providers ─────────────────────────────────────────────────────────
openai>=1.3.0
langchain>=1.2.0
langchain-core>=1.2.10
langchain-openai>=0.3.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.1.0

# ── AI/ML Framework ────────────────────────────────────────────────────────
crewai>=0.1.0

# ── Settings & Validation ──────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Utilities ──────────────────────────────────────────────────────────────
requests>=2.31.0
pyyaml>=6.0
cryptography==46.0.5
watchdog==6.0.0

# ── Streamlit (legacy app — DO NOT REMOVE) ─────────────────────────────────
streamlit==1.54.0

# ── FastAPI backend ────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# ── Async PostgreSQL ───────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# ── Database migrations ────────────────────────────────────────────────────
alembic>=1.13.0

# ── JWT validation ─────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0

# ── Dev/Testing ────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
